-Fixed property changes from 1.044
-Fixed missing methods from 1.044
-Changed cube grid managers over to new manager manager
-Cleaned up compile warnings

+++ SEModAPIInternal/API/Entity/Sector/SectorObject/CubeGrid/CubeBlock/ReactorEntity.cs
@@ -109,7 +109,7 @@ namespace SEModAPIInternal.API.Entity.Sector.SectorObject.CubeGrid.CubeBlock
 				float fuelMass = 0;
 				foreach (var item in ObjectBuilder.Inventory.Items)
 				{
-					fuelMass += item.Amount;
+					fuelMass += (float)item.Amount;
 				}
 				return fuelMass;
 			}
